malicious url detectionV1
My project focuses on malicious URL detection, a crucial aspect of cybersecurity. It involves developing and implementing algorithms and techniques to identify and flag potentially harmful or malicious web addresses. The primary goal is to protect users and systems from cyber threats such as phishing attacks, malware downloads, and fraud.